Implement session creation (login), also using changesets
We're using the changeset for casting/scrubbing values and validation
that it provides. And this behaves as expected. The issue is that
because the changeset is never sent through Workshop.Repo, its :action
is never set, and apprently Phoenix.HTML's form_for helper cares about
the :action when deciding whether to render error messages.

In the case of login, errors are generated as expected on the changeset
when login fails but they're never displayed on the page because no
database action was taken/attempted on the changeset.

This makes it harder to use changesets for non-database-backed actions
while integrating nicely with Phoneix's built-in form helpers.