Use two different changesets for new and create actions
There's no reason to run validations on the new action, where the
purpose is to simply cast any preset values and render the form.

By default, Phoenix.HTML will not render the error messages because we
haven't tried to persist the user yet, but that behavior doesn't feel
right and it's a waste of cycles to run the validations just to have
them wholly ignored.